requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argcomplete==3.5.3
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
datamodel-code-generator==0.26.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==33.3.1
genson==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ijson==3.3.0
inflect==5.6.2
isort==5.13.2
Jinja2==3.1.5
joblib==1.4.2
jsf==0.11.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
label-studio-sdk==1.0.8
lxml==5.3.0
MarkupSafe==3.0.2
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
nltk==3.9.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
rpds-py==0.22.3
rstr==3.2.2
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.3.0
wrapt==1.17.2
xmljson==0.2.1
